requirements.txt
flask
pandas
python-binance
numpy